% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_network.R
\name{cor_network}
\alias{cor_network}
\title{Correlation Network}
\usage{
cor_network(data, tax.aggregate = "Phylum", tax.add = NULL,
  tax.class = NULL, tax.empty = "best", threshold.cor = 0.8,
  threshold.count = 10)
}
\arguments{
\item{data}{(required) A phyloseq object.}

\item{tax.aggregate}{The taxonomic level that the data should be aggregated to (default: "Phylum")}

\item{tax.add}{Additional taxonomic levels to display for each entry, e.g. "Phylum" (default: "none").}

\item{tax.class}{Converts a specific phyla to class level instead, e.g. "p__Proteobacteria" (default: "none").}

\item{tax.empty}{Either "remove" OTUs without taxonomic information, add "best" classification or add the "OTU" name (default: "best").}

\item{threshold.cor}{Threshold value for the correlation (default: 0.8).}

\item{threshold.count}{Threshold value for abundance reads (default: 10).}
}
\description{
This function allows you to build a network based on correlations
}
\keyword{correlation}
\keyword{network}

